requirements.txt
cyrtranslit==0.4
Django==2.2.19
django-tinymce4-lite==1.8.0
jsmin==2.2.2
mysqlclient==2.0.3
pycodestyle==2.6.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
